+++ src/samudra/data/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2026 Samudra Authors
+#
+# SPDX-License-Identifier: Apache-2.0
